feat : Board Entity 수정
