+++ src/main/java/com/wanted/breadcrumbs/entity/board/Board.java
@@ -1,0 +1,17 @@
+package com.wanted.breadcrumbs.entity.board;
+
+import lombok.*;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@EqualsAndHashCode
+@ToString
+public class Board {
+
+    private Long id;
+    private String title;
+    private String contents;
+    private Long parentId;
+}
